Refactor Dependency Plugin
