+++ antibytes-dependency/src/main/kotlin/tech/antibytes/gradle/dependency/Version.kt
@@ -6,263 +6,38 @@
 
 package tech.antibytes.gradle.dependency
 
+import tech.antibytes.gradle.dependency.version.Android
+import tech.antibytes.gradle.dependency.version.Google
+import tech.antibytes.gradle.dependency.version.Gradle
+import tech.antibytes.gradle.dependency.version.Jvm
+import tech.antibytes.gradle.dependency.version.Kotlin
+import tech.antibytes.gradle.dependency.version.Npm
+import tech.antibytes.gradle.dependency.version.Square
+
 object Version {
 
     val gradle = Gradle
 
-    object Gradle {
-        /**
-         * [KTLint](https://github.com/pinterest/ktlint/releases)
-         */
-        const val ktLint = "0.47.1"
-    }
-
     // Kotlin
     val kotlin = Kotlin
 
-    object Kotlin {
-        /**
-         * [Kotlin](https://github.com/JetBrains/kotlin)
-         */
-        const val language = "1.7.10"
-
-        const val stdlib = language
-
-        /**
-         * [Kotlin Coroutines](https://github.com/Kotlin/kotlinx.coroutines)
-         */
-        const val coroutines = "1.6.4"
-
-        /**
-         * [Koin](https://github.com/InsertKoinIO/koin)
-         */
-        const val koin = "3.2.0"
-
-        /**
-         * [Ktor](https://github.com/ktorio/ktor)
-         */
-        const val ktor = "2.1.1"
-
-        /**
-         * [Kotlin Serialization](https://github.com/Kotlin/kotlinx.serialization)
-         */
-        const val serialization = "1.4.0"
-
-        /**
-         * [Kotlin DateTime](https://github.com/Kotlin/kotlinx-datetime)
-         */
-        const val dateTime = "0.4.0"
-
-        /**
-         * [UUID](https://github.com/benasher44/uuid)
-         */
-        const val uuid = "0.5.0"
-
-        /**
-         * [Stately](https://github.com/touchlab/Stately)
-         */
-        const val stately = "1.2.3"
-
-        /**
-         * (AtomicFu)(https://github.com/Kotlin/kotlinx.atomicfu)
-         */
-        const val atomicFu = "0.18.3"
-
-        val test = Test
-
-        object Test {
-            /**
-             * [mockk](http://mockk.io)
-             */
-            const val mockk = "1.12.8"
-        }
-    }
-
     val square = Square
-
-    object Square {
-        /**
-         * (Okio)(https://square.github.io/okio/)
-         */
-        const val okio = "3.2.0"
-
-        /**
-         * [SQLDelight](https://github.com/cashapp/sqldelight/)
-         */
-        const val sqldelight = "1.5.3"
-
-        /**
-         * [OkHttp](https://square.github.io/okhttp/)
-         */
-        const val okhttp = "4.10.0"
-    }
 
     val android = Android
 
-    object Android {
-        /**
-         * [Android Gradle Plugin](https://developer.android.com/studio/releases/gradle-plugin)
-         */
-        const val androidGradlePlugin = "7.2.2"
-
-        /**
-         * [Android Desugaring explained](https://developer.android.com/studio/write/java8-support)
-         * [Android Desugaring](https://github.com/google/desugar_jdk_libs)
-         */
-        const val desugar = "1.1.5"
-
-        // AndroidX
-        /**
-         * [AndroidX](https://developer.android.com/jetpack/androidx)
-         */
-
-        /**
-         * [AndroidX Core](https://developer.android.com/kotlin/ktx)
-         */
-        val ktx = Ktx
-
-        /**
-         * [AndroidX Tech](https://androidx.tech/)
-         */
-        object Ktx {
-            const val core = "1.9.0"
-
-            /**
-             * [Android Tech](https://androidx.tech/artifacts/activity/activity-ktx/)
-             */
-            const val activity = "1.5.1"
-            const val collections = "1.2.0"
-            const val lifecycle = "2.5.1"
-            const val fragment = "1.5.2"
-            const val livedata = "2.5.1"
-            const val navigation = "2.5.2"
-            const val palette = "1.0.0"
-            const val pageing = "3.1.1"
-            const val viewmodel = "2.5.1"
-            const val workmanager = "2.7.1"
-
-            /**
-             * [Annotation](https://developer.android.com/jetpack/androidx/releases/annotation)
-             */
-            const val annotation = "1.4.0"
-        }
-
-        /**
-         * [AppCompact](https://developer.android.com/jetpack/androidx/releases/appcompat)
-         */
-        const val appCompat = "1.5.1"
-
-        /**
-         * [Constrain Layout](https://developer.android.com/jetpack/androidx/releases/constraintlayout)
-         */
-        const val constraintLayout = "2.1.4"
-
-        // Material
-        /**
-         * [Material Android](https://github.com/material-components/material-components-android)
-         */
-        const val material = "1.7.0-rc01"
-
-        /**
-         * [Jetpack Compose](https://developer.android.com/jetpack/androidx/releases/compose)
-         * [Release Notes](https://developer.android.com/jetpack/androidx/versions/all-channel)
-         * [Compatibility Notes](https://developer.android.com/jetpack/androidx/releases/compose-kotlin)
-         */
-
-        val compose = Compose
-
-        object Compose {
-            const val core = "1.2.1"
-            const val compiler = "1.3.1"
-            const val foundation = "1.2.1"
-            const val material = "1.2.1"
-            const val runtime = "1.2.1"
-            const val ui = "1.2.1"
-            const val constraintLayout = "1.0.1"
-            const val activity = "1.5.1"
-            const val viewmodel = "2.5.1"
-            const val paging = "1.0.0-alpha16"
-        }
-
-        val test = Test
-
-        object Test {
-            /**
-             * [Android Testing](https://developer.android.com/testing)
-             */
-            /**
-             * [Android Testing Releases](https://developer.android.com/jetpack/androidx/releases/test)
-             */
-            const val test = "1.4.0"
-            const val espresso = "3.4.0"
-            const val orchestrator = "1.4.1"
-            const val junit = "1.1.3"
-
-            /**
-             * [Android UITest Automation explained](https://developer.android.com/training/testing/ui-testing/uiautomator-testing)
-             * [Android UITest Automation](https://androidx.tech/artifacts/test.uiautomator/uiautomator/)
-             */
-            const val uiAutomator = "2.2.0"
-
-            /**
-             * [Robolectric](https://github.com/robolectric/robolectric)
-             */
-            const val robolectric = "4.8.2"
-        }
-
-        /**
-         * [Coil](https://github.com/coil-kt/coil)
-         */
-        const val coil = "2.2.1"
-    }
-
     val jvm = Jvm
-
-    object Jvm {
-        /**
-         * [Log4J](http://www.slf4j.org/)
-         */
-        const val slf4j = "1.7.36"
-
-        val test = Test
-
-        object Test {
-            /**
-             * [JUnit](https://github.com/junit-team/junit5/)
-             */
-            const val junit = "5.9.0"
-
-            /**
-             * [JUnit](https://github.com/junit-team/junit4/)
-             */
-            const val junit4 = "4.13.2"
-        }
-    }
 
     val google = Google
 
-    object Google {
-        /**
-         * [Google Hilt](https://developer.android.com/training/dependency-injection/hilt-android)
-         */
-        const val hilt = "2.43.2"
-
-        /**
-         * [Google Hilt Compose](https://developer.android.com/jetpack/androidx/releases/hilt)
-         */
-        const val hiltCompose = "1.0.0"
-    }
-
     val js = JS
+
+    val npm = Npm
 
     object JS {
         /**
          * [KotlinNodeJS](https://github.com/Kotlin/kotlinx-nodejs)
          */
         const val nodeJs = "0.0.7"
-
-        val test = Test
-
-        object Test
+        const val node = "18.9.0"
     }
 }
