+++ antibytes-dependency/src/main/kotlin/tech/antibytes/gradle/dependency/version/Google.kt
@@ -1,0 +1,19 @@
+/*
+ * Copyright (c) 2022 Matthias Geisler (bitPogo) / All rights reserved.
+ *
+ * Use of this source code is governed by Apache License, Version 2.0
+ */
+
+package tech.antibytes.gradle.dependency.version
+
+object Google {
+    /**
+     * [Google Hilt](https://developer.android.com/training/dependency-injection/hilt-android)
+     */
+    const val hilt = "2.43.2"
+
+    /**
+     * [Google Hilt Compose](https://developer.android.com/jetpack/androidx/releases/hilt)
+     */
+    const val hiltCompose = "1.0.0"
+}
