+++ antibytes-dependency/src/main/kotlin/tech/antibytes/gradle/dependency/version/Gradle.kt
@@ -1,0 +1,14 @@
+/*
+ * Copyright (c) 2022 Matthias Geisler (bitPogo) / All rights reserved.
+ *
+ * Use of this source code is governed by Apache License, Version 2.0
+ */
+
+package tech.antibytes.gradle.dependency.version
+
+object Gradle {
+    /**
+     * [KTLint](https://github.com/pinterest/ktlint/releases)
+     */
+    const val ktLint = "0.47.1"
+}
