+++ antibytes-dependency/src/main/kotlin/tech/antibytes/gradle/dependency/version/Npm.kt
@@ -4,8 +4,8 @@
  * Use of this source code is governed by Apache License, Version 2.0
  */
 
-package tech.antibytes.gradle.dependency
+package tech.antibytes.gradle.dependency.version
 
-object JSVersion {
+object Npm {
     const val node = "18.7.0"
 }
