+++ antibytes-dependency/src/main/kotlin/tech/antibytes/gradle/dependency/version/Square.kt
@@ -1,0 +1,24 @@
+/*
+ * Copyright (c) 2022 Matthias Geisler (bitPogo) / All rights reserved.
+ *
+ * Use of this source code is governed by Apache License, Version 2.0
+ */
+
+package tech.antibytes.gradle.dependency.version
+
+object Square {
+    /**
+     * (Okio)(https://square.github.io/okio/)
+     */
+    const val okio = "3.2.0"
+
+    /**
+     * [SQLDelight](https://github.com/cashapp/sqldelight/)
+     */
+    const val sqldelight = "1.5.3"
+
+    /**
+     * [OkHttp](https://square.github.io/okhttp/)
+     */
+    const val okhttp = "4.10.0"
+}
