+++ gradle/wrapper/gradle-wrapper.properties
@@ -1,9 +1,3 @@
-#
-# Copyright (c) 2022 Matthias Geisler (bitPogo) / All rights reserved.
-#
-# Use of this source code is governed by Apache License, Version 2.0
-#
-
 distributionBase=GRADLE_USER_HOME
 distributionPath=wrapper/dists
 distributionUrl=https\://services.gradle.org/distributions/gradle-7.5.1-all.zip
